[REF] Refactor ensure_git_repo validation using a decorator
