+++ src/devrules/cli_commands/commit.py
@@ -4,9 +4,10 @@ from typing import Any, Callable, Dict, Optional
 import typer
 
 from devrules.config import load_config
-from devrules.core.git_service import ensure_git_repo, get_current_branch, get_current_issue_number
+from devrules.core.git_service import get_current_branch, get_current_issue_number
 from devrules.messages import commit as msg
 from devrules.utils import gum
+from devrules.utils.decorators import ensure_git_repo
 from devrules.utils.typer import add_typer_block_message
 from devrules.validators.commit import validate_commit
 from devrules.validators.forbidden_files import (
@@ -84,6 +85,7 @@ def _build_commit_with_typer(tags: list[str]) -> Optional[str]:
 
 def register(app: typer.Typer) -> Dict[str, Callable[..., Any]]:
     @app.command()
+    @ensure_git_repo()
     def check_commit(
         file: str,
         config_file: Optional[str] = typer.Option(
@@ -107,9 +109,10 @@ def register(app: typer.Typer) -> Dict[str, Callable[..., Any]]:
             raise typer.Exit(code=0)
         else:
             typer.secho(f"✘ {result_message}", fg=typer.colors.RED)
-        raise typer.Exit(code=1)
+            raise typer.Exit(code=1)
 
     @app.command()
+    @ensure_git_repo()
     def commit(
         message: str,
         config_file: Optional[str] = typer.Option(
@@ -152,7 +155,6 @@ def register(app: typer.Typer) -> Dict[str, Callable[..., Any]]:
             typer.secho(f"\n✘ {result_message}", fg=typer.colors.RED)
             raise typer.Exit(code=1)
 
-        ensure_git_repo()
         current_branch = get_current_branch()
 
         # Check if current branch is protected (e.g., staging branches for merging)
@@ -228,6 +230,7 @@ def register(app: typer.Typer) -> Dict[str, Callable[..., Any]]:
             raise typer.Exit(code=1) from e
 
     @app.command()
+    @ensure_git_repo()
     def icommit(
         config_file: Optional[str] = typer.Option(
             None, "--config", "-c", help="Path to config file"
@@ -240,7 +243,6 @@ def register(app: typer.Typer) -> Dict[str, Callable[..., Any]]:
         import subprocess
 
         config = load_config(config_file)
-        ensure_git_repo()
 
         typer.secho("Checking commit requirements...", fg=typer.colors.BLUE)
